==> 06_beta_diversity.Rmd <==
# Beta diversity

```{r load_data_beta, comment="", message=FALSE, warning=FALSE, eval=TRUE}
load("../data/all_wombats/data.Rdata")
load("../data/all_wombats/beta.Rdata")
load("../data/all_wombats/beta_nmds.Rdata")
library(rmarkdown)
library(R.utils)
library(knitr)
library(tidyverse)
library(devtools)
library(tinytable)
library(broom)
library(broom.mixed)
library(janitor)
library(glue)

# For tree handling
library(ape)
library(phyloseq)
library(phytools)

# For plotting
library(plotly)
library(ggplot2)
library(ggrepel)
library(ggpubr)
library(ggnewscale)
library(gridExtra)
library(ggtreeExtra)
library(ggtree)
library(ggh4x)

# For statistics
library(spaa)
library(vegan)
library(Rtsne)
library(geiger)
library(hilldiv2)
library(distillR)
library(ANCOMBC)
library(lme4)
library(matrixStats)
```

#### Calculate beta Hill div

```{r beta_div, comment="", message=FALSE, warning=FALSE, eval=FALSE}
beta_q0n <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 0)

beta_q1n <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 1)

beta_q2n <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 2)

genome_counts_filt <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0))%>%
  rownames_to_column(., "genome")

genome_tree <- keep.tip(genome_tree, tip=genome_counts_filt$genome)

beta_q0p <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 0, tree = genome_tree)

beta_q1p <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 1, tree = genome_tree)

beta_q2p <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 2, tree = genome_tree)

genome_counts_filt <- genome_counts_filt[genome_counts_filt$genome %in% rownames(genome_gifts),]
genome_counts_filt <- genome_counts_filt %>%
  remove_rownames() %>% 
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0))%>%
  rownames_to_column(., "genome")

genome_gifts1 <- genome_gifts[rownames(genome_gifts) %in% genome_counts_filt$genome,]
genome_gifts1 <- genome_gifts1[, colSums(genome_gifts1 != 0) > 0]

dist <- genome_gifts1 %>%
  to.elements(., GIFT_db) %>%
  traits2dist(., method = "gower")

beta_q0f <- genome_counts_filt %>%
  #remove_rownames() %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 0, dist = dist)

beta_q1f <- genome_counts_filt %>%
  #remove_rownames() %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 1, dist = dist)

beta_q2f <- genome_counts_filt %>%
  #remove_rownames() %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  hillpair(., q = 2, dist = dist)

```
#### Save outputs
```{r save_beta, comment="", message=FALSE,echo=FALSE,warning=FALSE, eval=FALSE}
save(beta_q0n, 
     beta_q1n,
     beta_q2n,
     beta_q0p,
     beta_q1p,
     beta_q2p,
     beta_q0f,
     beta_q1f,
     beta_q2f,
     genome_counts_filt,
     sample_metadata,
     genome_metadata,
     genome_gifts,
     genome_gifts1,
     genome_tree,
     file = "../data/all_wombats/beta.Rdata")
```

### Permanova analysis
```{r}
set.seed(1234)
```

#### Richness
```{r permanova_q0, comment="", message=FALSE, warning=FALSE}
sample_metadata_row <- column_to_rownames(sample_metadata, "sample") 
sample_metadata_row <- sample_metadata_row[labels(beta_q0n$S), ]
# sample_metadata_trim <- sample_metadata %>%
#   filter(sample != "EHI01806" & sample != "EHI01812")

betadisper(beta_q0n$S, sample_metadata_row$host_species) %>% permutest(., pairwise = TRUE) 

adonis2(beta_q0n$S ~ host_species, 
        data = sample_metadata %>% arrange(match(sample,labels(beta_q0n$S))), 
        permutations = 999) %>%
        broom::tidy() %>%
        tt()
```

#### Neutral
```{r permanova_wombats_q1n, comment="", message=FALSE, warning=FALSE}
betadisper(beta_q1n$S, sample_metadata_row$host_species) %>% permutest(., pairwise = TRUE) 

adonis2(beta_q1n$S ~ host_species, 
        data = sample_metadata %>% arrange(match(sample,labels(beta_q1n$S))), 
        permutations = 999) %>%
        broom::tidy() %>%
        tt()
```

## Simpsons

```{r permanova_wombats_q2n, comment="", message=FALSE, warning=FALSE}
betadisper(beta_q2n$S, sample_metadata_row$host_species) %>% permutest(., pairwise = TRUE) 

adonis2(beta_q2n$S ~ host_species, 
        data = sample_metadata %>% arrange(match(sample,labels(beta_q2n$S))), 
        permutations = 999) %>%
        broom::tidy() %>%
        tt()
```



#### Phylogenetic
```{r permanova_wombats_q1p, comment="", message=FALSE, warning=FALSE}
betadisper(beta_q1p$S, sample_metadata_row$host_species) %>% permutest(., pairwise = TRUE)
adonis2(beta_q1p$S ~ host_species,
        data = sample_metadata %>% arrange(match(sample,labels(beta_q1p$S))),
        permutations = 999) %>%
        broom::tidy() %>%
        tt()
```

#### Functional
```{r permanova_q1F, comment="", message=FALSE, warning=FALSE}
betadisper(beta_q1f$S, sample_metadata_row$host_species) %>% permutest(., pairwise = TRUE)
adonis2(beta_q1f$S ~ host_species,
        data = sample_metadata %>% arrange(match(sample,labels(beta_q1f$S))),
        permutations = 999) %>%
        broom::tidy() %>%
        tt()
```

### Beta diversity plots

#### Richness

```{r beta_div_nmds_richness_plot_median, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q0n$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    ) 

) 
```
#### Neutral (host species)

```{r beta_div_nmds_neutral_plot_median_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q1n$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_order,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

) 
```

```{r beta_div_nmds_simpson_plot_median_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q2n$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_order,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

)

```


#### Neutral

```{r beta_div_neutral, comment="", echo=FALSE, message=FALSE, warning=FALSE}

ggplotly(

beta_q1n$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    ) 
# +geom_text_repel(aes(label = individual), size=3)

) 
```
#### Neutral (host species)

```{r beta_div_nmds_neutral_plot_median_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q1n$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_order,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

)

```

```{r beta_div_nmds_functional_plot_median_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}
ggplotly(

beta_q1f$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_order,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

)
```
# Simpsons for functional

```{r beta_div_nmds_functional_plot_simpson_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q2f$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_order,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

)

```
# Richness for functional

```{r  beta_div_nmds_functional_plot_richness_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q0f$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_order,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

)

```



```{r beta_div_nmds_phylogenetic_plot_species, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q1p$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_species,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

)

```

# Let's lookat phylogenetic richness

```{r beta_div_nmds_phylogenetic_plot_richness, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q0p$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_species,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

)

```

# phylogenetic diversity according to q2 (Simpson's)

```{r beta_div_nmds_phylogenetic_plot_simpsons, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

ggplotly(

beta_q2p$S %>%
  vegan::metaMDS(., trymax = 500, k = 2, trace=0) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  dplyr::left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(x = NMDS1, y = NMDS2, color = host_species, shape = host_species,
             label = sample, label1 = singlem_fraction)) +
    geom_point(size = 4) +
    #   stat_ellipse(aes(color = beta_q1n_nmds$Groups))+
    geom_segment(aes(x = x_cen, y = y_cen, xend = NMDS1, yend = NMDS2), alpha = 0.9, show.legend = FALSE) +
    theme_classic() +
    theme(
      axis.text.x = element_text(size = 12),
      axis.text.y = element_text(size = 12),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(face = "bold", size = 12),
      panel.background = element_blank(),
      axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
      legend.text = element_text(size = 12),
      legend.title = element_text(size = 14),
      legend.position = "right", legend.box = "vertical"
    )

)

```


# Let'stry a PCoA of the beta diversity and see what that looks like. 

```{r PCoA_richness_bq0n, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

pcoa <- cmdscale(beta_q0n$S, eig = TRUE, add = TRUE)
positions <- pcoa$points
colnames(positions) <- c("pcoa1","pcoa2")

percent_explained <- 100 * pcoa$eig / sum(pcoa$eig)

pretty_pe <- round(percent_explained[1:2], digits = 1)

labs <- c(glue("PCo 1 ({pretty_pe[1]}%)"),
          glue("PCo 1 ({pretty_pe[2]}%)"))

positions %>%
  as_tibble(rownames = "sample") %>%
  inner_join(sample_metadata, pcoa, by = "sample") %>%
ggplot(aes(x=pcoa1, y=pcoa2, colour = host_species)) +
  geom_point(shape=16, alpha=0.7, size = 3) +
  labs(x=labs[1], y =labs[2]) +
  theme_minimal()
  
tibble(pe = percent_explained,
       axis = 1:length(percent_explained)) %>%
  ggplot(aes(x=axis, y=pe)) +
  geom_line() +
  coord_cartesian(xlim=c(1,5))


```

# Neutral
```{r PCoA_neutral_bq1n, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

pcoa <- cmdscale(beta_q1n$S, eig = TRUE, add = TRUE)
positions <- pcoa$points
colnames(positions) <- c("pcoa1","pcoa2")

percent_explained <- 100 * pcoa$eig / sum(pcoa$eig)

pretty_pe <- round(percent_explained[1:2], digits = 1)

labs <- c(glue("PCo 1 ({pretty_pe[1]}%)"),
          glue("PCo 1 ({pretty_pe[2]}%)"))

positions %>%
  as_tibble(rownames = "sample") %>%
  inner_join(sample_metadata, pcoa, by = "sample") %>%
ggplot(aes(x=pcoa1, y=pcoa2, colour = host_species)) +
  geom_point(shape=16, alpha=0.7, size = 3) +
  labs(x=labs[1], y =labs[2]) +
  theme_minimal()
  
tibble(pe = percent_explained,
       axis = 1:length(percent_explained)) %>%
  ggplot(aes(x=axis, y=pe)) +
  geom_line() +
  coord_cartesian(xlim=c(1,5))

```

```{r PCoA_phylogenetic_bq1p, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

pcoa <- cmdscale(beta_q1p$S, eig = TRUE, add = TRUE)
positions <- pcoa$points
colnames(positions) <- c("pcoa1","pcoa2")

percent_explained <- 100 * pcoa$eig / sum(pcoa$eig)

pretty_pe <- round(percent_explained[1:2], digits = 1)

labs <- c(glue("PCo 1 ({pretty_pe[1]}%)"),
          glue("PCo 1 ({pretty_pe[2]}%)"))

positions %>%
  as_tibble(rownames = "sample") %>%
  inner_join(sample_metadata, pcoa, by = "sample") %>%
ggplot(aes(x=pcoa1, y=pcoa2, colour = host_species)) +
  geom_point(shape=16, alpha=0.7, size = 3) +
  labs(x=labs[1], y =labs[2]) +
  theme_minimal()
  
tibble(pe = percent_explained,
       axis = 1:length(percent_explained)) %>%
  ggplot(aes(x=axis, y=pe)) +
  geom_line() +
  coord_cartesian(xlim=c(1,5))
```

```{r PCoA_functional_bq1f, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

pcoa <- cmdscale(beta_q1f$S, eig = TRUE, add = TRUE)
positions <- pcoa$points
colnames(positions) <- c("pcoa1","pcoa2")

percent_explained <- 100 * pcoa$eig / sum(pcoa$eig)

pretty_pe <- round(percent_explained[1:2], digits = 1)

labs <- c(glue("PCo 1 ({pretty_pe[1]}%)"),
          glue("PCo 1 ({pretty_pe[2]}%)"))

positions %>%
  as_tibble(rownames = "sample") %>%
  inner_join(sample_metadata, pcoa, by = "sample") %>%
ggplot(aes(x=pcoa1, y=pcoa2, colour = host_species)) +
  geom_point(shape=16, alpha=0.7, size = 3) +
  labs(x=labs[1], y =labs[2]) +
  theme_minimal()
  
tibble(pe = percent_explained,
       axis = 1:length(percent_explained)) %>%
  ggplot(aes(x=axis, y=pe)) +
  geom_line() +
  coord_cartesian(xlim=c(1,5))
```


# lets try a different method such as Bray-Curtis

```{r bray_cutris_test, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

# looking for low count MAGs or samples
molten_table_filt <- molten_table %>%
  group_by(genome) %>%
  mutate(total = sum(read_count)) %>%
  arrange(total) %>%
  select(-covered_fraction, -MAG_length)

molten_table_filt %>%
  ggplot(aes(x=total))+
  geom_histogram()

# Creating a matrix for Bray-Curtis NMDS
genome_counts_filt_matrix <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0)) %>%
  t() %>%
  as.matrix()

dist_bray <- vegdist(genome_counts_filt_matrix, method = "bray")
set.seed(1234)
nmds_bray <- metaMDS(dist_bray, trymax = 500, k = 2, trace=0)


```

```{r, save_nmds, comment="", message=FALSE,echo=FALSE,warning=FALSE, eval=FALSE }

save(genome_counts_filt_matrix,
     dist_bray,
     nmds_bray,
     file = "../data/all_wombats/beta_nmds.Rdata")
```

```{r bray_cutris_nmds, comment="", message=FALSE, warning=FALSE, fig.height=7, fig.width=10, fig.fullwidth=TRUE}

scores(nmds_bray) %>%
  as_tibble(., rownames = "sample") %>%
  left_join(sample_metadata, by = join_by(sample == sample)) %>%
  ggplot(aes(x=NMDS1, y=NMDS2, colour = host_species)) +
  geom_point(shape=16, alpha=0.7, size = 3) +
  theme_classic()
  

```
