==> 08_differential_abundance_functional.Rmd <==

# Functional differences

```{r load_data_func,comment="", echo=FALSE, message=FALSE, warning=FALSE}
load("../data/all_wombats/genome_gifts_raw.Rdata")
load("../data/all_wombats/data.Rdata")
#load("data/ancombc_filtered_functional.Rdata")
library(forcats)
library(rmarkdown)
library(R.utils)
library(knitr)
library(tidyverse)
library(devtools)
library(tinytable)
library(broom)
library(broom.mixed)
library(janitor)

# For tree handling
library(ape)
library(phyloseq)
library(phytools)

# For plotting
library(plotly)
library(ggplot2)
library(ggrepel)
library(ggpubr)
library(ggnewscale)
library(gridExtra)
library(ggtreeExtra)
library(ggtree)
library(ggh4x)
library(patchwork)

# For statistics
library(spaa)
library(vegan)
library(Rtsne)
library(geiger)
library(hilldiv2)
library(distillR)
library(ANCOMBC)
library(lme4)
library(matrixStats)
```

```{r gift_analysis, comment="", echo=FALSE, message=FALSE, warning=FALSE}
genome_counts_filt <- genome_counts_filt[genome_counts_filt$genome %in% rownames(genome_gifts_raw),]
genome_counts_filt <- genome_counts_filt %>%
  column_to_rownames(., "genome") %>%
  filter(rowSums(. != 0, na.rm = TRUE) > 0) %>%
  select_if(~!all(. == 0))%>%
  rownames_to_column(., "genome")

genome_gifts <- genome_gifts_raw[rownames(genome_gifts_raw) %in% genome_counts_filt$genome,]
genome_gifts <- genome_gifts[, colSums(genome_gifts != 0) > 0]

#Aggregate bundle-level GIFTs into the compound level
GIFTs_elements <- to.elements(genome_gifts,GIFT_db)
GIFTs_elements_filtered <- GIFTs_elements[rownames(GIFTs_elements) %in% genome_counts_filt$genome,]
GIFTs_elements_filtered <- as.data.frame(GIFTs_elements_filtered) %>% 
  select_if(~ !is.numeric(.) || sum(.) != 0)

#Aggregate element-level GIFTs into the function level
GIFTs_functions <- to.functions(GIFTs_elements_filtered,GIFT_db)

#Aggregate function-level GIFTs into overall Biosynthesis, Degradation and Structural GIFTs
GIFTs_domains <- to.domains(GIFTs_functions,GIFT_db)

#Get community-weighed average GIFTs per sample
genome_counts_row <- genome_counts_filt %>%
  mutate_at(vars(-genome),~./sum(.)) %>% 
  column_to_rownames(., "genome") 
# genome_counts_row <- rownames_to_column(genome_counts_row, "genome")
GIFTs_elements_community <- to.community(GIFTs_elements_filtered,genome_counts_row,GIFT_db)
GIFTs_functions_community <- to.community(GIFTs_functions,genome_counts_row,GIFT_db)
GIFTs_domains_community <- to.community(GIFTs_domains,genome_counts_row,GIFT_db)
```

```{r phylum_elem_capacity, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=12, fig.width=10, fig.fullwidth=TRUE}
func_heatmap_phylum <- 
  GIFTs_elements_filtered %>% 
  rownames_to_column("genome") %>%
  left_join(genome_metadata[,1:8], by = join_by(genome == genome)) %>% group_by(phylum)  %>%
  summarize(across(where(is.numeric), mean, na.rm = TRUE))%>%
    pivot_longer(!phylum,names_to="trait",values_to="gift") %>%
    mutate(functionid = substr(trait, 1, 3)) %>%
    mutate(trait = case_when(
      trait %in% GIFT_db$Code_element ~ GIFT_db$Element[match(trait, GIFT_db$Code_element)],
      TRUE ~ trait
    )) %>%
    mutate(functionid = case_when(
      functionid %in% GIFT_db$Code_function ~ GIFT_db$Function[match(functionid, GIFT_db$Code_function)],
      TRUE ~ functionid
    )) %>%
    mutate(trait=factor(trait,levels=unique(GIFT_db$Element))) %>%
    mutate(functionid=factor(functionid,levels=unique(GIFT_db$Function))) %>%
    ggplot(aes(x=phylum,y=trait,fill=gift)) +
        geom_tile(colour="white", linewidth=0.2)+
        scale_fill_gradientn(colours=rev(c("#d53e4f", "#f46d43", "#fdae61", "#fee08b", "#e6f598", "#abdda4", "#ddf1da")))+
        facet_grid(functionid ~ ., scales="free",space="free") +
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
              axis.text.y = element_text(size=6),
              strip.text.y = element_text(angle = 0),
              ) +
        labs(y="Traits",x="Phylum",fill="GIFT")

```

## Functional capacity of the structural MAGs
```{r struct_zero_capwild, comment="", echo=FALSE, message=FALSE, warning=FALSE}
samples <- sample_metadata %>% 
  dplyr::select(sample) %>% pull()

nhnw_samples <- sample_metadata %>% 
  dplyr::select(sample, host_species) %>%
  dplyr::filter(., host_species == "Lasiorhinus krefftii") %>%
  dplyr::select(sample) %>%
  pull()

shnw_samples <- sample_metadata %>% 
  dplyr::select(sample, host_species) %>%
  dplyr::filter(., host_species == "Lasiorhinus latifrons") %>%
  dplyr::select(sample) %>%
  pull()

bnw_samples <- sample_metadata %>% 
  dplyr::select(sample, host_species) %>%
  dplyr::filter(., host_species == "Vombatus ursinus") %>%
  dplyr::select(sample) %>%
  pull()

structural_zeros <- genome_counts_filt %>% 
   rowwise() %>% #compute for each row (genome)
   mutate(all_zeros_all = all(c_across(all_of(samples)) == 0)) %>% # set true if all samples in pre have zeros
   mutate(average_all = mean(c_across(all_of(samples)), na.rm = TRUE)) %>% # get average genome counts across pre
   mutate(all_zeros_nhnw = all(c_across(all_of(nhnw_samples)) == 0)) %>% 
   mutate(average_nhnw = mean(c_across(all_of(nhnw_samples)), na.rm = TRUE)) %>% # get average genome counts across wild
   mutate(all_zeros_shnw = all(c_across(all_of(shnw_samples)) == 0)) %>% 
   mutate(average_shnw = mean(c_across(all_of(shnw_samples)), na.rm = TRUE)) %>%
   mutate(all_zeros_bnw = all(c_across(all_of(bnw_samples)) == 0)) %>% 
   mutate(average_bnw = mean(c_across(all_of(bnw_samples)), na.rm = TRUE)) %>%
   filter(all_zeros_nhnw == TRUE || all_zeros_shnw==TRUE || all_zeros_bnw==TRUE) %>%  # filter only genomes with structural zeros
   mutate(present = case_when(
      all_zeros_nhnw & !all_zeros_shnw & !all_zeros_bnw ~ "nhnw",
      !all_zeros_nhnw & all_zeros_shnw & !all_zeros_bnw ~ "shnw",
      !all_zeros_nhnw & !all_zeros_shnw & all_zeros_bnw ~ "bnw",
      !all_zeros_nhnw & !all_zeros_shnw & !all_zeros_bnw ~ "none",
      TRUE ~ NA_character_
    )) %>%
   mutate(average = case_when(present == "nhnw" ~ average_nhnw,
                              present == "shnw" ~ average_shnw,
                              present == "bnw" ~ average_bnw)) %>%
   dplyr::select(genome, present, average) %>%
   left_join(genome_metadata, by=join_by(genome==genome)) %>%
   arrange(present,-average)
structural_zeros
struct_mag <- structural_zeros %>% 
  filter(present %in% c("nhnw","shnw","bnw")) %>% 
  select(genome) %>% 
  pull()
struct_mag_meta <- structural_zeros %>% 
  filter(present %in% c("nhnw","shnw","bnw"))
```
```{r structural_functional_capacity2, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=12,  fig.fullwidth=TRUE}
GIFTs_elements_filtered %>% 
  rownames_to_column("genome") %>%
  filter(genome %in% struct_mag)  %>% 
    pivot_longer(!genome,names_to="trait",values_to="gift") %>% 
    left_join(struct_mag_meta, by = join_by(genome == genome))%>%
    mutate(functionid = substr(trait, 1, 3)) %>%
    mutate(trait = case_when(
      trait %in% GIFT_db$Code_element ~ GIFT_db$Element[match(trait, GIFT_db$Code_element)],
      TRUE ~ trait
    )) %>%
    mutate(functionid = case_when(
      functionid %in% GIFT_db$Code_function ~ GIFT_db$Function[match(functionid, GIFT_db$Code_function)],
      TRUE ~ functionid
    )) %>%
    mutate(trait=factor(trait,levels=unique(GIFT_db$Element))) %>%
    mutate(functionid=factor(functionid,levels=unique(GIFT_db$Function))) %>%
    ggplot(aes(x=genome,y=trait,fill=gift)) +
        geom_tile(colour="white", linewidth=0.2)+
        scale_fill_gradientn(colours=rev(c("#d53e4f", "#f46d43", "#fdae61", "#fee08b", "#e6f598", "#abdda4", "#ddf1da")))+
        facet_grid(functionid ~ present, scales="free",space="free") +
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1, size=4),
              axis.text.y = element_text(size=4),
              strip.text.y = element_text(angle = 0)
              ) +
        labs(y="Traits",x="Samples",fill="GIFT")

```

## Phylum MCI
```{r mci_mean, echo=FALSE, message=FALSE, warning=FALSE}
GIFTs_elements_filtered  %>% 
  rownames_to_column("genome")%>% 
  left_join(genome_metadata[,1:8], by = join_by(genome == genome)) %>% group_by(phylum)  %>%  # Group by phylum
  summarize(across(where(is.numeric), mean, na.rm = TRUE)) %>%  # Calculate the mean per group
  mutate(
    row_mean = rowMeans(across(where(is.numeric)), na.rm = TRUE),  # Calculate row means
    row_sd = apply(across(where(is.numeric)), 1, sd, na.rm = TRUE),
    row_median = apply(across(where(is.numeric)), 1, median, na.rm = TRUE),
    row_max = apply(across(where(is.numeric)), 1, max, na.rm = TRUE),
    row_min = apply(across(where(is.numeric)), 1, min, na.rm = TRUE)# Calculate row standard deviations
  ) %>% 
  select(phylum, row_mean, row_sd, row_median, row_max, row_min) %>% 
  arrange(-row_mean)
```

```{r phylum_elem_values, comment="", echo=FALSE, message=FALSE, warning=FALSE}
GIFTs_elements_filtered %>% 
  rownames_to_column("genome") %>%
  left_join(genome_metadata[,1:8], by = join_by(genome == genome)) %>% group_by(phylum)  %>%
  summarize(across(where(is.numeric), mean, na.rm = TRUE))%>%
    pivot_longer(!phylum,names_to="trait",values_to="gift") %>%
    mutate(functionid = substr(trait, 1, 3)) %>%
    mutate(trait = case_when(
      trait %in% GIFT_db$Code_element ~ GIFT_db$Element[match(trait, GIFT_db$Code_element)],
      TRUE ~ trait
    )) %>%
    mutate(functionid = case_when(
      functionid %in% GIFT_db$Code_function ~ GIFT_db$Function[match(functionid, GIFT_db$Code_function)],
      TRUE ~ functionid
    )) %>%
    mutate(trait=factor(trait,levels=unique(GIFT_db$Element))) %>%
    mutate(functionid=factor(functionid,levels=unique(GIFT_db$Function))) %>% 
  group_by(phylum, functionid) %>% 
  summarize(median=median(gift))%>%
  pivot_wider(names_from = functionid, values_from = median)
  
```

## Groups MCI
```{r gitfs_functional_species, echo=TRUE,results=TRUE}
GIFTs_functions_community %>%
  rowMeans() %>%
  as_tibble(., rownames = "sample") %>%
  left_join(sample_metadata, by = join_by(sample == sample)) %>%
  group_by(host_species) %>%
  summarise(MCI = mean(value), sd = sd(value)) %>%
  tt()


 
GIFTs_functions_community %>%
  rowMeans() %>%
  as_tibble(., rownames = "sample") %>%
  left_join(sample_metadata, by = join_by(sample == sample)) %>%
  filter(host_species != "Vombatus ursinus") %>%
  filter(reserve != "RUNR") %>%
  group_by(season) %>%
  summarise(MCI = mean(value), sd = sd(value)) %>%
  tt()

```

***Shapiro test***
```{r}
MCI <- GIFTs_functions_community %>%
  rowMeans() %>%
  as_tibble(., rownames = "sample") %>%
  left_join(sample_metadata, by = join_by(sample == sample)) %>%
  filter(host_species != "Vombatus ursinus") %>%
  filter(sample != "B478_23_EFNP_NB") %>%
  filter(sample != "B11_23_RUNR") %>%
  filter(sample != "B19_2_23_RUNR") %>%
  filter(sample != "FS1_1_23_RUNR")
  

MCI_season <- MCI %>%
  filter(reserve != "RUNR")

shapiro.test(MCI_season$value)
res.aov.season <- aov(value ~ season, data=MCI_season)
summary(res.aov.season)

shapiro.test(MCI$value)
res.aov <- aov(value ~ host_species, data=MCI)
summary(res.aov)

t.test(value ~ host_species, data = MCI)
t.test(value ~ season, data = MCI_season)
```

## Community level plots
```{r comunity_elem_plot, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
func_heatmap_elem_season <- 
  GIFTs_elements_community %>%
    as.data.frame() %>%
    rownames_to_column(var="sample")%>%
    pivot_longer(!sample,names_to="trait",values_to="gift") %>%
    left_join(sample_metadata, by = join_by(sample == sample)) %>%
    filter(!host_species =="Vombatus ursinus")%>%
    filter(!host_species =="Lasiorhinus latifrons")%>%
    #filter(!host_species =="Lasiorhinus krefftii")%>%
    filter(!reserve == "RUNR") %>%
    mutate(functionid = substr(trait, 1, 3)) %>%
    mutate(trait = case_when(
      trait %in% GIFT_db$Code_element ~ GIFT_db$Element[match(trait, GIFT_db$Code_element)],
      TRUE ~ trait
    )) %>%
    mutate(functionid = case_when(
      functionid %in% GIFT_db$Code_function ~ GIFT_db$Function[match(functionid, GIFT_db$Code_function)],
      TRUE ~ functionid
    )) %>%
    mutate(trait=factor(trait,levels=unique(GIFT_db$Element))) %>%
    mutate(functionid=factor(functionid,levels=unique(GIFT_db$Function))) %>%
    ggplot(aes(x=sample,y=trait,fill=gift)) +
        geom_tile(colour="white", linewidth=0.2)+
        scale_fill_gradientn(colours=rev(c("#d53e4f", "#f46d43", "#fdae61", "#fee08b", "#e6f598", "#abdda4", "#ddf1da")))+
        facet_grid(functionid ~ host_species, scales="free",space="free") +
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
              axis.text.y = element_text(size=6),
              strip.text.y = element_text(angle = 0)
              ) +
        labs(y="Traits",x="Samples",fill="GIFT")
```

```{r comunity_funct_plot, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
unique_funct_db<- GIFT_db[c(3,4,5)] %>% 
  distinct(Code_function, .keep_all = TRUE)

func_heatmap_comunity_spring <- 
  GIFTs_functions_community %>%
   t() %>%
  as.data.frame() %>%
  mutate_if(is.character, as.numeric)  %>%
  rownames_to_column(., "Code_function")  %>%
  left_join(.,unique_funct_db[c(1,3)],by = join_by(Code_function == Code_function))  %>%
  select(-Code_function) %>%
  column_to_rownames(., "Function")%>%
   t()  %>%
    as.data.frame() %>%
    rownames_to_column(var="sample") %>%
    pivot_longer(!sample,names_to="trait",values_to="gift") %>%
    left_join(sample_metadata, by = join_by(sample == sample))%>%
    #filter(!host_species =="Lasiorhinus krefftii") %>%
    filter(!reserve == "RUNR") %>%
    filter(!host_species =="Lasiorhinus latifrons") %>%
    filter(!host_species =="Vombatus ursinus") %>%
    #filter(!season == "spring") %>%
    filter(!season == "winter") %>%
    ggplot(aes(x=trait,y=sample,fill=gift)) +
        geom_tile(colour="white", linewidth=0.2)+
        scale_fill_gradientn(colours=rev(c("#d53e4f", "#f46d43", "#fdae61", "#fee08b", "#e6f598", "#abdda4", "#ddf1da")))+
        facet_grid(host_species ~ ., scales="free",space="free")+
  theme(#axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1, size = 10),
        axis.text.x  = element_blank(),
        axis.ticks.x = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        strip.background = element_blank(),
        strip.text = element_text(size = 12, color="black",face="bold"),
        axis.title = element_text(size = 12, face = "bold"),
        #axis.title.y = element_text(margin = margin(t = 0, r = 0, b = 0, l = 0)),
        axis.title.y = element_blank(),
        axis.title.x = element_blank(),
        #axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0)),
        panel.background= element_blank()
              ) +
        labs(x="Function", y="Sample",fill="GIFT")
```

## combined heatmap
```{r comunity_funct_plot_combined, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}

heat_map_combined <- func_heatmap_comunity_NHNW + func_heatmap_comunity_SHNW + func_heatmap_comunity_BNW + plot_layout(ncol = 1)

```

```{r comunity_funct_plot_combined, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}

heat_map_combined_season <- func_heatmap_comunity_spring + func_heatmap_comunity_winter + plot_layout(ncol = 1)
```


## combining heatmaps for figure
```{r combined_heatmap, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}

heatmap_combined <- func_heatmap_phylum + func_heatmap_comunity + plot_layout(guides = 'collect')


```




## Kruskal Wallis comparison
### NHNW vs SHNW vs BNW 
#### Community elements differences:
```{r comunity_elem, comment="", echo=FALSE, message=FALSE, warning=FALSE}
element_gift <- GIFTs_elements_community %>% 
  as.data.frame() %>% 
  rownames_to_column(., "sample") %>% 
  right_join(sample_metadata[c(1,3)], by="sample")

element_gift_season <- GIFTs_elements_community %>% 
  as.data.frame() %>% 
  rownames_to_column(., "sample") %>% 
  right_join(sample_metadata[c(1,3,23,24)], by="sample") %>%
  filter(reserve == 'EFNP') %>%
  select(-host_species, -reserve)
```

```{r commun_wilcox_elem, comment="", echo=FALSE, message=FALSE, warning=FALSE}
uniqueGIFT_db<- unique(GIFT_db[c(2,4,5,6)]) %>% unite("Function",Function:Element, sep= "_", remove=FALSE)

significant_elements <- element_gift %>%
    filter(host_species != "Vombatus ursinus") %>%
    filter(sample != "B478_23_EFNP_NB") %>%
    filter(sample != "B11_23_RUNR") %>%
    filter(sample != "B19_2_23_RUNR") %>%
    filter(sample != "FS1_1_23_RUNR") %>%
    pivot_longer(-c(sample,host_species), names_to = "trait", values_to = "value") %>%
    group_by(trait) %>%
    summarise(p_value = kruskal.test(value ~ host_species)$p.value) %>%
    mutate(p_adjust=p.adjust(p_value, method="BH")) %>%
    filter(p_adjust < 0.05)%>%
  rownames_to_column(., "Elements") %>%
  select(-Elements)

significant_elements_season <- element_gift_season %>%
    pivot_longer(-c(sample,season), names_to = "trait", values_to = "value") %>%
    group_by(trait) %>%
    summarise(p_value = kruskal.test(value ~ season)$p.value) %>%
    mutate(p_adjust=p.adjust(p_value, method="BH")) %>%
    filter(p_adjust < 0.05)%>%
  rownames_to_column(., "Elements") %>%
  select(-Elements)

element_gift_t <- element_gift  %>% 
  t() %>%
  row_to_names(row_number = 1) %>%
  as.data.frame() %>%
  mutate_if(is.character, as.numeric)  %>%
  rownames_to_column(., "trait")

element_gift_filt <- subset(element_gift_t, trait %in% significant_elements$trait) %>% 
  t() %>%
  row_to_names(row_number = 1) %>%
  as.data.frame() %>%
  mutate_if(is.character, as.numeric)  %>%
  rownames_to_column(., "sample")%>% 
  left_join(., sample_metadata[c(1,3)], by = join_by(sample == sample)) %>%
  filter(host_species != "Vombatus ursinus") %>%
  filter(sample != "B478_23_EFNP_NB") %>%
  filter(sample != "B11_23_RUNR") %>%
  filter(sample != "B19_2_23_RUNR") %>%
  filter(sample != "FS1_1_23_RUNR")

element_gift_filt_seaason <- subset(element_gift_t, trait %in% significant_elements_season$trait) %>% 
  t() %>%
  row_to_names(row_number = 1) %>%
  as.data.frame() %>%
  mutate_if(is.character, as.numeric)  %>%
  rownames_to_column(., "sample")%>% 
  left_join(., sample_metadata[c(1,3, 23,24)], by = join_by(sample == sample)) %>%
  filter(reserve == "EFNP") %>%
  filter(sample != "B478_23_EFNP_NB") %>%
  select(-host_species, -reserve)

difference_table <- element_gift_filt %>%
  select(-sample) %>%
  group_by(host_species) %>%
  summarise(across(everything(), mean)) %>%
  t() %>%
  row_to_names(row_number = 1) %>%
  as.data.frame() %>%
  mutate_if(is.character, as.numeric) %>%
  rownames_to_column(., "Elements") %>%
  left_join(.,uniqueGIFT_db[c(1,3,4)],by = join_by(Elements == Code_element)) %>%
  arrange(Function) %>%
  mutate(Difference_N_S = `Lasiorhinus krefftii` - `Lasiorhinus latifrons`) %>%
  mutate(group_color_N_S = ifelse(Difference_N_S <0, "shnw","nhnw"))
  
difference_table_season <- element_gift_filt_seaason %>%
  select(-sample) %>%
  group_by(season) %>%
  summarise(across(everything(), mean)) %>%
  t() %>%
  row_to_names(row_number = 1) %>%
  as.data.frame() %>%
  mutate_if(is.character, as.numeric) %>%
  rownames_to_column(., "Elements") %>%
  left_join(.,uniqueGIFT_db[c(1,3,4)],by = join_by(Elements == Code_element)) %>%
  arrange(Function) %>%
  mutate(Difference_W_S = `winter` - `spring`) %>%
  mutate(group_color_W_S = ifelse(Difference_W_S <0, "spring","winter"))
```

```{r log_fold_calc, comment="", echo=FALSE, message=FALSE, warning=FALSE}
means_gift <- element_gift_filt %>% 
  select(-host_species) %>%
  pivot_longer(!sample, names_to = "elements", values_to = "abundance") %>%
  left_join(sample_metadata, by=join_by(sample==sample)) %>%
  group_by(host_species, elements) %>%
  summarise(mean=mean(abundance))

means_gift_season <- element_gift_filt_seaason %>% 
  select(-season) %>%
  pivot_longer(!sample, names_to = "elements", values_to = "abundance") %>%
  left_join(sample_metadata, by=join_by(sample==sample)) %>%
  group_by(season, elements) %>%
  summarise(mean=mean(abundance))

log_fold <- means_gift %>%
  group_by(elements) %>%
  summarise(
    logfc_nhnw_shnw = log2(mean[host_species == "Lasiorhinus krefftii"] / mean[host_species == "Lasiorhinus latifrons"]))

log_fold_season <- means_gift_season %>%
  group_by(elements) %>%
  summarise(
    logfc_winter_spring = log2(mean[season == "winter"] / mean[season == "spring"]))
```


```{r plot3, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=5, fig.width=10, fig.fullwidth=TRUE}
uniqueGIFT <- unique(GIFT_db[c(2,3,4,5,6)])

code_function2 <- difference_table %>%
  left_join(uniqueGIFT[c(1:3)], by=join_by(Elements==Code_element))

code_function2_season <- difference_table_season %>%
  left_join(uniqueGIFT[c(1:3)], by=join_by(Elements==Code_element))

unique_codes<-unique(code_function2$Code_function)

gift_colors <- read_tsv("../data/all_wombats/gift_colors.tsv") %>% 
  filter(Code_function %in% unique_codes)%>% 
  mutate(legend=str_c(Code_function," - ",Function))
```

```{r final_plot_nhnw_shnw, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=12, fig.fullwidth=TRUE}
pos <- position_jitter(width = 0.25, height = 0.25, seed = 4321)

nhnw_shnw_log_fold <- 
  code_function2 %>%
  left_join(significant_elements, by=join_by(Elements==trait)) %>%
  left_join(log_fold, by=join_by(Elements==elements)) %>% 
  left_join(gift_colors, by=join_by(Code_function==Code_function)) %>% 
  ggplot(., aes(x = logfc_nhnw_shnw, y = -log(p_adjust), color=legend, size=abs(Difference_N_S))) +
  geom_jitter(position = pos)+
  geom_vline(xintercept=0) +
  scale_color_manual(values = gift_colors$Color)+
  theme_classic()+
  labs(size="Mean difference (abs)", color="Functional trait")+
  labs(x = "Log-fold change", y="-Log adjusted p-value") +
  geom_text_repel(aes(label = Element), min.segment.length = 0.4, size=2.5, max.overlaps = Inf, 
                  position = pos) +
  theme(
    legend.title = element_text(size = 7), 
    legend.text = element_text(size = 6))
```


```{r final_plot_season, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=12, fig.fullwidth=TRUE}
pos <- position_jitter(width = 0.1, height = 0.1, seed = 1)

season_log_fold <- 
  code_function2_season %>%
  left_join(significant_elements_season, by=join_by(Elements==trait)) %>%
  left_join(log_fold_season, by=join_by(Elements==elements)) %>% 
  left_join(gift_colors, by=join_by(Code_function==Code_function)) %>% 
  ggplot(., aes(x = logfc_winter_spring, y = -log(p_adjust), color=legend, size=abs(Difference_W_S))) +
  geom_jitter(position = pos)+
  geom_vline(xintercept=0) +
  scale_color_manual(values = gift_colors$Color)+
  theme_classic()+
  labs(size="Mean difference (abs)", color="Functional trait")+
  labs(x = "Log-fold change", y="-Log adjusted p-value") +
  geom_text_repel(aes(label = Element), min.segment.length = 0.1, size=2.5, max.overlaps = Inf, position = pos) +
  theme(
    legend.title = element_text(size = 7),
    legend.text = element_text(size = 6))
```

```{r commun_wilcox_elem_plot_nhnw_shnw, comment="", echo=FALSE, message=FALSE, warning=FALSE,fig.fullwidth=TRUE, fig.fullheight =TRUE, eval=FALSE}
 #nhnw_shnw_diff <- 
  difference_table %>%
  ggplot(aes(x=forcats::fct_reorder(Function,Difference_N_S), y=Difference_N_S, fill=group_color_N_S)) + 
  geom_col() +
#  geom_point(size=4) + 
  scale_fill_manual(values=c("#e5bd5b", "#6b7398")) + 
  geom_hline(yintercept=0) + 
  coord_flip()+
  theme(axis.text = element_text(size = 6),
        axis.title = element_text(size = 12),
        legend.position = "right", 
        legend.title = element_blank(),
        panel.background = element_blank(),
          panel.grid.major = element_line(size = 0.15, linetype = 'solid',
                                colour = "grey"))+
  xlab("Function") + 
  ylab("Mean difference")
```

```{r commun_wilcox_elem_plot_nhnw_bnw, comment="", echo=FALSE, message=FALSE, warning=FALSE,fig.fullwidth=TRUE, fig.fullheight =TRUE, eval=FALSE}
nhnw_bnw_diff <- difference_table %>%
  ggplot(aes(x=forcats::fct_reorder(Function,Difference_B_N), y=Difference_B_N, fill=group_color_B_N)) + 
  geom_col() +
#  geom_point(size=4) + 
  scale_fill_manual(values=c("#e5bd5b", "#6b7398")) + 
  geom_hline(yintercept=0) + 
  coord_flip()+
  theme(axis.text = element_text(size = 6),
        axis.title = element_text(size = 12),
        legend.position = "right", 
        legend.title = element_blank(),
        panel.background = element_blank(),
          panel.grid.major = element_line(size = 0.15, linetype = 'solid',
                                colour = "grey"))+
  xlab("Function") + 
  ylab("Mean difference")
```

```{r commun_wilcox_elem_plot_shnw_bnw, comment="", echo=FALSE, message=FALSE, warning=FALSE,fig.fullwidth=TRUE, fig.fullheight =TRUE, eval=FALSE}
shnw_bnw_diff <- difference_table %>%
  ggplot(aes(x=forcats::fct_reorder(Function,Difference_B_S), y=Difference_B_S, fill=group_color_B_S)) + 
  geom_col() +
#  geom_point(size=4) + 
  scale_fill_manual(values=c("#e5bd5b", "#6b7398")) + 
  geom_hline(yintercept=0) + 
  coord_flip()+
  theme(axis.text = element_text(size = 6),
        axis.title = element_text(size = 12),
        legend.position = "right", 
        legend.title = element_blank(),
        panel.background = element_blank(),
          panel.grid.major = element_line(size = 0.15, linetype = 'solid',
                                colour = "grey"))+
  xlab("Function") + 
  ylab("Mean difference")
```




```{r sugar_plot2, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=6, fig.width=12, fig.fullwidth=TRUE}
sugar_function <- GIFT_db %>% 
  filter(Code_function=="D03") %>% 
  select(Code_element) %>% 
  pull()

GIFTs_elements_community %>% 
  as.data.frame() %>% 
  select(any_of(sugar_function)) %>% 
  rownames_to_column(., "sample") %>% 
  left_join(sample_metadata[c(1,21)], by=join_by(sample==sample)) %>% 
  select(-sample) %>% 
  pivot_longer(!diet,names_to="trait",values_to="gift") %>% 
  left_join(uniqueGIFT_db, by=join_by(trait==Code_element)) %>% 
  ggplot(aes(x=diet, y=gift, color=diet)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=diet_colors)+
  scale_fill_manual(values=diet_colors) +
  scale_x_discrete(labels=c("Wild" = "Wild", "Pre_grass" = "Captive-born")) +
facet_grid(.~Element,  scales="free_x")+
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10, angle = 45, hjust = 1),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )
```

```{r sugar_plot, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=6, fig.width=12, fig.fullwidth=TRUE}
poly_function <- GIFT_db %>% 
  filter(Code_function=="D02") %>% 
  select(Code_element) %>% 
  pull()

GIFTs_elements_community %>% 
  as.data.frame() %>% 
  select(any_of(poly_function)) %>% 
  rownames_to_column(., "sample") %>% 
  left_join(sample_metadata[c(1,21)], by=join_by(sample==sample)) %>% 
  select(-sample) %>% 
  pivot_longer(!diet,names_to="trait",values_to="gift") %>% 
  left_join(uniqueGIFT_db, by=join_by(trait==Code_element)) %>% 
  ggplot(aes(x=diet, y=gift, color=diet)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=diet_colors)+
  scale_fill_manual(values=diet_colors) +
  scale_x_discrete(labels=c("Wild" = "Wild", "Pre_grass" = "Captive-born")) +
facet_grid(.~Element,  scales="free_x")+
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10, angle = 45, hjust = 1),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )
```


### host species functional diffreences

#### Community elements differences:

***ANCOMBC***
```{r phyloseq_ele_host_species, comment="", echo=FALSE, message=FALSE, warning=FALSE}
sample_info_tab_phy <- sample_metadata%>%
  column_to_rownames(var="sample") %>%
  sample_data()

count_phy <- GIFTs_elements_community %>% 
  as.data.frame() %>%
  rownames_to_column("sample") %>%
  filter(sample %in% rownames(sample_info_tab_phy)) %>%
  column_to_rownames("sample") %>% 
  select(which(!colSums(., na.rm=TRUE) %in% 0)) %>% 
  t() %>%
  otu_table(., taxa_are_rows=T)

uniqueGIFT_db<- unique(GIFT_db[c(2,4,5,6)]) %>% unite("Function",Function:Element, sep= "_", remove=FALSE)

TAX <- uniqueGIFT_db%>%
    remove_rownames()%>%
  column_to_rownames(var="Code_element")%>%
  as.matrix()%>%
  tax_table()

physeq_function = phyloseq(count_phy, TAX, sample_info_tab_phy)
  
```

```{r ancom_rand_elem_post_bnw_ref, comment="", echo=FALSE, message=FALSE, warning=FALSE, eval=FALSE}
set.seed(1234) #set seed for reproducibility
ancom_rand_output_element_bnw_ref = ancombc2(data = physeq_function, 
                  assay_name = "counts", 
                  tax_level = NULL, #change to agglomerate analysis to a higher taxonomic range
                  fix_formula = "group", #fixed variable(s)
                 # rand_formula = "(1|individual)",
                  p_adj_method = "holm", 
                  pseudo_sens = TRUE,
                  prv_cut = 0, 
                  lib_cut = 0, 
                  s0_perc = 0.05,
                  group = "group", 
                  struc_zero = TRUE, 
                  neg_lb = FALSE,
                  alpha = 0.05, 
                  n_cl = 2, 
                  verbose = TRUE,
                  global = FALSE, 
                  pairwise = FALSE, 
                  dunnet = FALSE, 
                  trend = FALSE,
                  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
                  em_control = list(tol = 1e-5, max_iter = 100),
                  lme_control = lme4::lmerControl(),
                  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100), 
                  trend_control = NULL)

```

```{r table_ele2_bnw_ref, comment="", echo=FALSE, message=FALSE, warning=FALSE}
tax <- data.frame(physeq_function@tax_table) %>%
  rownames_to_column(., "taxon")

ancombc_rand_elem_bnw_ref <- ancom_rand_output_element_bnw_ref$res %>%
  dplyr::select(taxon, lfc_groupNHNW, p_groupNHNW, lfc_groupSHNW, p_groupSHNW) %>%
  filter(p_groupNHNW < 0.05) %>%
  filter(p_groupSHNW < 0.05) %>%
  dplyr::arrange(p_groupNHNW) %>%
  merge(., tax, by="taxon") %>%
  dplyr::arrange(lfc_groupNHNW)

```

```{r ancom_rand_res_elem2_bnw_ref, echo=FALSE, comment="", message=FALSE, warning=FALSE}
ancombc_rand_table_bnw_ref <- ancom_rand_output_element_bnw_ref$res %>%
  dplyr::select(taxon, lfc_groupNHNW, p_groupNHNW, lfc_groupSHNW, p_groupSHNW) %>%
  filter(p_groupNHNW < 0.05) %>%
  filter(p_groupSHNW < 0.05) %>%
  dplyr::arrange(p_groupNHNW) %>%
  merge(., tax, by="taxon")
```

## NHNW vs. BNW (BNW ref)

```{r ancombc_rand_plot_elem2_nhnw_bnw, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
func_ancombc_nhnw_bnw <- ancombc_rand_table_bnw_ref%>%
  mutate(Color = ifelse(lfc_groupNHNW <0, "bnw","nhnw")) %>%
  ggplot(aes(x=forcats::fct_reorder(Function,lfc_groupNHNW), y=lfc_groupNHNW, fill=Color)) + 
  geom_col() +
  scale_fill_manual(values=c("#e5bd5b", "#6b7398")) + 
  geom_hline(yintercept=0) + 
  coord_flip()+
  theme(axis.text = element_text(size = 10, face="bold.italic"),
        axis.title = element_text(size = 12),
        legend.position = "right", 
        legend.title = element_blank(),
        panel.background = element_blank(),
          panel.grid.major = element_line(size = 0.15, linetype = 'solid',
                                colour = "grey"))+
  xlab("Traits") + 
  ylab("log2FoldChange")
```

```{r ancombc_rand_plot_elem2_shnw_bnw, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
func_ancombc_shnw_bnw <- ancombc_rand_table_bnw_ref%>%
  mutate(Color = ifelse(lfc_groupSHNW <0, "bnw","shnw")) %>%
  ggplot(aes(x=forcats::fct_reorder(Function,lfc_groupNHNW), y=lfc_groupNHNW, fill=Color)) + 
  geom_col() +
  scale_fill_manual(values=c("#e5bd5b", "#6b7398")) + 
  geom_hline(yintercept=0) + 
  coord_flip()+
  theme(axis.text = element_text(size = 10, face="bold.italic"),
        axis.title = element_text(size = 12),
        legend.position = "right", 
        legend.title = element_blank(),
        panel.background = element_blank(),
          panel.grid.major = element_line(size = 0.15, linetype = 'solid',
                                colour = "grey"))+
  xlab("Traits") + 
  ylab("log2FoldChange")

```
## using NHNW as the ref for NHNW vs. SHNW comparison
```{r ancom_rand_elem_post_nhnw_ref, comment="", echo=FALSE, message=FALSE, warning=FALSE, eval=FALSE}

set.seed(1234) #set seed for reproducibility
ancom_rand_output_element_nhnw_ref = ancombc2(data = physeq_function, 
                  assay_name = "counts", 
                  tax_level = NULL, #change to agglomerate analysis to a higher taxonomic range
                  fix_formula = "host_species", #fixed variable(s)
                 # rand_formula = "(1|individual)",
                  p_adj_method = "holm", 
                  pseudo_sens = TRUE,
                  prv_cut = 0, 
                  lib_cut = 0, 
                  s0_perc = 0.05,
                  group = "host_species", 
                  struc_zero = TRUE, 
                  neg_lb = FALSE,
                  alpha = 0.05, 
                  n_cl = 2, 
                  verbose = TRUE,
                  global = FALSE, 
                  pairwise = FALSE, 
                  dunnet = FALSE, 
                  trend = FALSE,
                  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
                  em_control = list(tol = 1e-5, max_iter = 100),
                  lme_control = lme4::lmerControl(),
                  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100), 
                  trend_control = NULL)
```


```{r table_ele2_nhnw_ref, comment="", echo=FALSE, message=FALSE, warning=FALSE}
tax <- data.frame(physeq_function@tax_table) %>%
  rownames_to_column(., "taxon")

ancombc_rand_elem_nhnw_ref <- ancom_rand_output_element_nhnw_ref$res %>%
  dplyr::select(taxon, `lfc_host_speciesLasiorhinus latifrons`, `p_host_speciesLasiorhinus latifrons`, `lfc_host_speciesVombatus ursinus`, `p_host_speciesVombatus ursinus`) %>%
  filter(`p_host_speciesLasiorhinus latifrons` < 0.05) %>%
  filter(`p_host_speciesVombatus ursinus` < 0.05) %>%
  dplyr::arrange(`p_host_speciesLasiorhinus latifrons`) %>%
  merge(., tax, by="taxon") %>%
  dplyr::arrange(`lfc_host_speciesLasiorhinus latifrons`)
```



```{r ancom_rand_res_elem2_nhnw_ref, echo=FALSE, comment="", message=FALSE, warning=FALSE}
ancombc_rand_table_nhnw_ref <- ancom_rand_output_element_nhnw_ref$res %>%
  dplyr::select(taxon, `lfc_host_speciesLasiorhinus latifrons`, `p_host_speciesLasiorhinus latifrons`, `lfc_host_speciesVombatus ursinus`, `p_host_speciesVombatus ursinus`) %>%
  filter(`p_host_speciesLasiorhinus latifrons` < 0.05) %>%
  filter(`p_host_speciesVombatus ursinus` < 0.05) %>%
  dplyr::arrange(`p_host_speciesLasiorhinus latifrons`) %>%
  merge(., tax, by="taxon")
```


```{r ancombc_rand_plot_elem2_nhnw_shnw, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
func_ancombc_nhnw_shnw <- 
  ancombc_rand_table_nhnw_ref%>%
  mutate(Color = ifelse(`lfc_host_speciesLasiorhinus latifrons` <0, "Lasiorhinus krefftii","Lasiorhinus latifrons")) %>%
  ggplot(aes(x=forcats::fct_reorder(Function,`lfc_host_speciesLasiorhinus latifrons`), y=`lfc_host_speciesLasiorhinus latifrons`, fill=Color)) + 
  geom_col() +
  scale_fill_manual(values=c("#01665e", "#5ab4ac")) + 
  geom_hline(yintercept=0) + 
  coord_flip()+
  theme(axis.text = element_text(size = 6, face="bold.italic", angle = 20),
        axis.title = element_text(size = 12),
        legend.position = "right", 
        legend.title = element_blank(),
        panel.background = element_blank(),
        panel.grid.major = element_line(size = 0.15, linetype = 'solid',
                                colour = "grey"))+
  xlab("Traits") + 
  ylab("log Fold Change")

```

***Wilcoxon comparison***
```{r comunity_elem, comment="", echo=FALSE, message=FALSE, warning=FALSE}
element_gift <- GIFTs_elements_community %>% 
  as.data.frame() %>% 
  rownames_to_column(., "sample") %>% 
  left_join(sample_metadata[c(1,3)], by="sample")
```

```{r commun_wilcox_elem, comment="", echo=FALSE, message=FALSE, warning=FALSE}
uniqueGIFT_db<- unique(GIFT_db[c(2,4,5,6)]) %>% unite("Function",Function:Element, sep= "_", remove=FALSE)

element_gift %>%
  pivot_longer(-c(sample,host_species), names_to = "trait", values_to = "value") %>%
  group_by(trait) %>%
  mutate(p_value = kruskal.test(value ~ host_species)$p.value) %>%
  mutate(p_adjust=p.adjust(p_value, method="BH")) %>%
  filter(p_adjust < 0.05)%>%
  rownames_to_column(., "Elements") %>% 
  select(-Elements) %>%
  summarise(host_species, trait, p_value, p_adjust)
```

## Let's test function differences between season in NHNW @ EFNP

```{r phyloseq_ele_post_season, comment="", echo=FALSE, message=FALSE, warning=FALSE}
sample_info_tab_phy_season <- sample_metadata%>%
  column_to_rownames(var="sample") %>%
  filter(host_species == "Lasiorhinus krefftii") %>%
  filter(reserve == "EFNP") %>%
  sample_data()

count_phy_season <- GIFTs_elements_community %>% 
  as.data.frame() %>%
  rownames_to_column("sample") %>%
  filter(sample %in% rownames(sample_info_tab_phy_season)) %>%
  column_to_rownames("sample") %>% 
  select(which(!colSums(., na.rm=TRUE) %in% 0)) %>% 
  t() %>%
  otu_table(., taxa_are_rows=T)

uniqueGIFT_db<- unique(GIFT_db[c(2,4,5,6)]) %>% unite("Function",Function:Element, sep= "_", remove=FALSE)

TAX <- uniqueGIFT_db%>%
    remove_rownames()%>%
  column_to_rownames(var="Code_element")%>%
  as.matrix()%>%
  tax_table()

physeq_function_season = phyloseq(count_phy_season, TAX, sample_info_tab_phy_season)
```



```{r ancom_rand_elem_season_spring_ref, comment="", echo=FALSE, message=FALSE, warning=FALSE, eval=FALSE}
set.seed(1234) #set seed for reproducibility
ancom_rand_output_element_sping_ref = ancombc2(data = physeq_function_season, 
                  assay_name = "counts", 
                  tax_level = NULL, #change to agglomerate analysis to a higher taxonomic range
                  fix_formula = "season", #fixed variable(s)
                 # rand_formula = "(1|individual)",
                  p_adj_method = "holm", 
                  pseudo_sens = TRUE,
                  prv_cut = 0, 
                  lib_cut = 0, 
                  s0_perc = 0.05,
                  group = "season", 
                  struc_zero = TRUE, 
                  neg_lb = FALSE,
                  alpha = 0.05, 
                  n_cl = 2, 
                  verbose = TRUE,
                  global = FALSE, 
                  pairwise = FALSE, 
                  dunnet = FALSE, 
                  trend = FALSE,
                  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
                  em_control = list(tol = 1e-5, max_iter = 100),
                  lme_control = lme4::lmerControl(),
                  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100), 
                  trend_control = NULL)
```


```{r table_ele2_spring_ref, comment="", echo=FALSE, message=FALSE, warning=FALSE}
tax <- data.frame(physeq_function_season@tax_table) %>%
  rownames_to_column(., "taxon")

ancombc_rand_elem_spring_ref <- ancom_rand_output_element_sping_ref$res %>%
  dplyr::select(taxon, lfc_seasonwinter, p_seasonwinter) %>%
  filter(p_seasonwinter < 0.05) %>%
  dplyr::arrange(p_seasonwinter) %>%
  merge(., tax, by="taxon") %>%
  dplyr::arrange(lfc_seasonwinter)
```

```{r ancom_rand_res_elem2_spring_ref, echo=FALSE, comment="", message=FALSE, warning=FALSE}
ancombc_rand_table_spring_ref <- ancom_rand_output_element_sping_ref$res %>%
  dplyr::select(taxon, lfc_seasonwinter, p_seasonwinter) %>%
  filter(p_seasonwinter < 0.05) %>%
  dplyr::arrange(p_seasonwinter) %>%
  merge(., tax, by="taxon")
```

```{r ancombc_rand_plot_elem2_season, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
ancombc_func_season <- ancombc_rand_table_spring_ref%>%
  mutate(Color = ifelse(lfc_seasonwinter <0, "spring","winter")) %>%
  ggplot(aes(x=forcats::fct_reorder(Function,lfc_seasonwinter), y=lfc_seasonwinter, fill=Color)) + 
  geom_col() +
  scale_fill_manual(values=c("#fdae61", "#2c7bb6")) + 
  geom_hline(yintercept=0) + 
  coord_flip()+
  theme(axis.text = element_text(size = 6, face="bold.italic", angle = 20),
        axis.title = element_text(size = 12),
        legend.position = "right", 
        legend.title = element_blank(),
        panel.background = element_blank(),
          panel.grid.major = element_line(size = 0.15, linetype = 'solid',
                                colour = "grey"))+
  xlab(element_blank()) + 
  ylab("log Fold Change")
```


```{r}
ancombc_func_combined <- func_ancombc_nhnw_shnw + ancombc_func_season + plot_layout(guides = 'collect') + plot_annotation(tag_levels = 'A')
```


## now lets look at the functional differences between NHNW at reserves

```{r phyloseq_ele_post_reserve, comment="", echo=FALSE, message=FALSE, warning=FALSE}
sample_info_tab_phy_reserve <- sample_metadata%>%
  column_to_rownames(var="sample") %>%
  filter(host_species == "Lasiorhinus krefftii") %>%
  filter(season == "winter") %>%
  sample_data()

count_phy_reserve <- GIFTs_elements_community %>% 
  as.data.frame() %>%
  rownames_to_column("sample") %>%
  filter(sample %in% rownames(sample_info_tab_phy_reserve)) %>%
  column_to_rownames("sample") %>% 
  select(which(!colSums(., na.rm=TRUE) %in% 0)) %>% 
  t() %>%
  otu_table(., taxa_are_rows=T)

uniqueGIFT_db<- unique(GIFT_db[c(2,4,5,6)]) %>% unite("Function",Function:Element, sep= "_", remove=FALSE)

TAX <- uniqueGIFT_db%>%
    remove_rownames()%>%
  column_to_rownames(var="Code_element")%>%
  as.matrix()%>%
  tax_table()

physeq_function_reserve = phyloseq(count_phy_reserve, TAX, sample_info_tab_phy_reserve)



```

```{r ancom_rand_elem_reserve_efnp_ref, comment="", echo=FALSE, message=FALSE, warning=FALSE, eval=FALSE}
set.seed(1234) #set seed for reproducibility
ancom_rand_output_element_efnp_ref = ancombc2(data = physeq_function_reserve, 
                  assay_name = "counts", 
                  tax_level = NULL, #change to agglomerate analysis to a higher taxonomic range
                  fix_formula = "reserve", #fixed variable(s)
                 # rand_formula = "(1|individual)",
                  p_adj_method = "holm", 
                  pseudo_sens = TRUE,
                  prv_cut = 0, 
                  lib_cut = 0, 
                  s0_perc = 0.05,
                  group = NULL, 
                  struc_zero = FALSE, 
                  neg_lb = FALSE,
                  alpha = 0.05, 
                  n_cl = 2, 
                  verbose = TRUE,
                  global = FALSE, 
                  pairwise = FALSE, 
                  dunnet = FALSE, 
                  trend = FALSE,
                  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
                  em_control = list(tol = 1e-5, max_iter = 100),
                  lme_control = lme4::lmerControl(),
                  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100), 
                  trend_control = NULL)
```

```{r table_ele2_efnp_ref, comment="", echo=FALSE, message=FALSE, warning=FALSE}
tax <- data.frame(physeq_function_reserve@tax_table) %>%
  rownames_to_column(., "taxon")

ancombc_rand_elem_efnp_ref <- ancom_rand_output_element_efnp_ref$res %>%
  dplyr::select(taxon, lfc_reserveRUNR, p_reserveRUNR) %>%
  filter(p_reserveRUNR < 0.05) %>%
  dplyr::arrange(p_reserveRUNR) %>%
  merge(., tax, by="taxon") %>%
  dplyr::arrange(lfc_reserveRUNR)
```

```{r ancom_rand_res_elem2_spring_ref, echo=FALSE, comment="", message=FALSE, warning=FALSE}
ancombc_rand_table_efnp_ref <- ancom_rand_output_element_efnp_ref$res %>%
  dplyr::select(taxon, lfc_reserveRUNR, p_reserveRUNR) %>%
  filter(p_reserveRUNR < 0.05) %>%
  dplyr::arrange(p_reserveRUNR) %>%
  merge(., tax, by="taxon")

```

```{r ancombc_rand_plot_elem2_reserve, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
 ancombc_func_reserve <- ancombc_rand_table_efnp_ref%>%
  mutate(Color = ifelse(lfc_reserveRUNR <0, "EFNP","RUNR")) %>%
  ggplot(aes(x=forcats::fct_reorder(Function,lfc_reserveRUNR), y=lfc_reserveRUNR, fill=Color)) + 
  geom_col() +
  scale_fill_manual(values=c("#e5bd5b", "#6b7398")) + 
  geom_hline(yintercept=0) + 
  coord_flip()+
  theme(axis.text = element_text(size = 10, face="bold.italic"),
        axis.title = element_text(size = 12),
        legend.position = "right", 
        legend.title = element_blank(),
        panel.background = element_blank(),
          panel.grid.major = element_line(size = 0.15, linetype = 'solid',
                                colour = "grey"))+
  xlab("Traits") + 
  ylab("log2FoldChange")


```
