==> 10_differential_analysis_no_structural_zeroes.Rmd <==
---
title: "Differential Abundance No Structural zeroes"
author: "Colin"
date: "2025-07-30"
output: html_document
---

```{r load_data, comment="", echo=FALSE, message=FALSE, warning=FALSE, eval=FALSE}

load("../data/all_wombats/data.Rdata")

```


# Create a phyloseq object for ANCOMBC
```{r create_phyloseq_object, comment="", message=FALSE, warning=FALSE, eval=FALSE }

# phyloseq object considering structual zeros
phylo_samples <- sample_metadata %>%
  column_to_rownames("sample") %>%
  sample_data() # convert to phyloseq sample_data object

phylo_genome <- genome_counts_filt %>%
  column_to_rownames("genome") %>%
  mutate_all(~ replace(., . == 0, 0.00001)) %>% # add pseudo counts to avoid structural zero issues (note this approach can be improved!) %>%
  mutate_all(~./sum(.)) %>% #apply TSS nornalisation
  otu_table(., taxa_are_rows = TRUE)

phylo_taxonomy <- genome_metadata %>%
  filter(genome %in% rownames(phylo_genome)) %>% # remove structural zeros
  mutate(genome2 = genome) %>% # create a pseudo genome name column
  column_to_rownames("genome2") %>%
  dplyr::select(domain, phylum, class, order, family, genus, species, genome) %>% # add an additional taxonomic level to ensure genome-level analysis (as no all genomes have species-level taxonomic assignments. Otherwise, ANCOMBC2 aggregates analyses per species)
  as.matrix() %>%
  tax_table() # convert to phyloseq tax_table object

phyloseq <- phyloseq(phylo_genome, phylo_taxonomy, phylo_samples)


```

# Run ANCOMBC
```{r ancombc, comment="", message=FALSE, warning=FALSE, eval=FALSE}
differential_abundance <- ancombc2(
  data = phyloseq,
  assay_name = "counts",
  tax_level = NULL, # change to agglomerate analysis to a higher taxonomic range
  fix_formula = "host_species", # fixed variable(s)
  rand_formula = "1| region",
  p_adj_method = "holm",
  pseudo_sens = TRUE,
  prv_cut = 0,
  lib_cut = 0,
  s0_perc = 0.05,
  group = NULL,
  struc_zero = FALSE,
  neg_lb = FALSE,
  alpha = 0.05,
  n_cl = 2,
  verbose = TRUE,
  global = FALSE,
  pairwise = FALSE,
  dunnet = FALSE,
  trend = FALSE,
  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
  em_control = list(tol = 1e-5, max_iter = 100),
  lme_control = lme4::lmerControl(),
  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100),
  trend_control = NULL
)

# Save differential abundance to data object
# save(sample_metadata, 
#      genome_metadata, 
# #     read_counts, 
# #     genome_counts, 
#      molten_table,
#      genome_counts_filt, 
#      genome_tree,
#      genome_gifts,
#      phylum_colors,
#      physeq,
#      physeq_clr,
#      species_colours,
#      differential_abundance,
#      file = "../data/all_wombats/data.Rdata")
```
